+++ lib-multisrc/madara/build.gradle.kts
@@ -6,4 +6,5 @@ baseVersionCode = 34
 
 dependencies {
     api(project(":lib:cryptoaes"))
+    api(project(":lib:i18n"))
 }
